Created very rudimentary login screen routes
