+++ app/App.js
@@ -1,59 +1,33 @@
 import "react-native-gesture-handler";
-import React from "react";
-import { NavigationContainer, DefaultTheme } from "@react-navigation/native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import React, { useState } from "react";
+import { View, Button } from "react-native";
+import { NavigationContainer } from "@react-navigation/native";
 
-import Icon from "react-native-vector-icons/FontAwesome";
+import MainTabNavigation from "./screens/MainTabNavigation";
+import Login from "./screens/Login";
+
 import theme from "./theme";
 
-// Screens
-import Home from "./screens/Home";
-import Parks from "./screens/Parks";
-import Profile from "./screens/Profile";
+const App = () => {
+    const [isSignedIn, setIsSignedIn] = useState(false);
 
-const Tab = createBottomTabNavigator();
+    const toggleSignIn = () => {
+        alert(isSignedIn);
+        setIsSignedIn(true);
+    };
 
-const MainTabNavigation = () => (
-    <Tab.Navigator
-        initialRouteName="Home"
-        screenOptions={({ route }) => ({
-            tabBarIcon: ({ focused, color, size }) => {
-                let iconName;
-
-                switch (route.name) {
-                    case "Home":
-                        iconName = "paw";
-                        break;
-                    case "Parks":
-                        iconName = "tree";
-                        break;
-                    case "Profile":
-                        iconName = "user";
-                        break;
-                    default:
-                        iconName = "question";
-                        break;
-                }
-
-                // You can return any component that you like here!
-                return <Icon name={iconName} size={size} color={color} />;
-            },
-        })}
-        tabBarOptions={{
-            activeTintColor: theme.colors.primary,
-            inactiveTintColor: "#5E503F",
-        }}
-    >
-        <Tab.Screen name="Parks" component={Parks} />
-        <Tab.Screen name="Home" component={Home} />
-        <Tab.Screen name="Profile" component={Profile} />
-    </Tab.Navigator>
-);
-
-const App = () => (
-    <NavigationContainer theme={theme}>
-        <MainTabNavigation />
-    </NavigationContainer>
-);
+    if (isSignedIn) {
+        return (
+            <NavigationContainer theme={theme}>
+                <MainTabNavigation theme={theme} />
+            </NavigationContainer>
+        );
+    }
+    return (
+        <View>
+            <Button title="Login" onPress={toggleSignIn} />
+        </View>
+    );
+};
 
 export default App;
