+++ app/screens/Login/Login.js
@@ -1,0 +1,6 @@
+import React from "react";
+import { Text } from "react-native";
+
+const Login = () => <Text>Login works!</Text>;
+
+export default Login;
