+++ app/screens/MainTabNavigation/index.js
@@ -1,0 +1,3 @@
+import MainTabNavigation from "./MainTabNavigation";
+
+export default MainTabNavigation;
