+++ app/screens/Profile/Profile.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { View, Text } from "react-native";
+import { View } from "react-native";
 import { Input } from "react-native-elements";
 import { createStackNavigator } from "@react-navigation/stack";
